estate-server/backend/app_backend/serializers.py
from rest_framework import serializers
from .models import CustomUser

class CustomUserSerializer(serializers.ModelSerializer):
    class Meta:
        model = CustomUser
        fields = ('id', 'username', 'email', 'first_name', 'last_name', 'role', 'preference', 'is_staff', 'is_active', 'is_superuser')
        read_only_fields = ('id', 'is_staff', 'is_active', 'is_superuser')

class UserLoginSerializer(serializers.Serializer):
    username = serializers.CharField()
    password = serializers.CharField(write_only=True)