estate-server/requirements.txt
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
psycopg2==2.9.6
sqlparse==0.4.3
